more modularization

+++ src/main.rs
@@ -11,48 +11,39 @@ use std::env;
 use std::process::Command;
 // use std::path::{Path, PathBuf};
 
-use nickel::{Nickel, HttpRouter, Mountable, StaticFilesHandler};
+use nickel::{Nickel, HttpRouter, Mountable, StaticFilesHandler, QueryString};
 
 use acled::{AllCountries, Event, EventTrunc, Country, CountryPageData, MainPageData,
-            CountryNameLink};
+            CountryNameLink, YearPageData};
 
 use postgres::{Connection, TlsMode};
 use postgres::params::{UserInfo, ConnectParams, ConnectTarget};
 // use postgres::types::ToSql;
 
-// use acled::AcledServer;
-
-
-fn contains_name(inp_vec: &Vec<Country>, inp_name: String) -> bool {
-    if inp_vec.is_empty() || inp_name.is_empty() {
-        false
-    }
-    else {
-        for elem in inp_vec {
-            if elem.name == inp_name {
-                return true;
-            }
-        }
-        false
-    }
-}
-
-fn get_country_by_link(inp_vec: &Vec<Country>, inp_link: String) -> Option<Country> {
-    if inp_vec.is_empty() || inp_link.is_empty() {
-        None
-    }
-    else {
-        for elem in inp_vec {
-            if elem.link == inp_link {
-                let n_elem = elem.clone();
-                return Some(n_elem);
-            }
-        }
-        None
-    }
-}
+use acled::AcledServer;
 
 fn main() {
+    // Some server testing
+    let mut test_server = AcledServer::new();
+    test_server.init();
+    let test_db_name = "acled_scratch";
+
+    let test_params = ConnectParams {
+        target: ConnectTarget::Tcp("localhost".to_owned()),
+        port: None,
+        user: Some(UserInfo {
+            user: "michael".to_owned(),
+            password: None,
+        }),
+        database: Some(test_db_name.to_owned()),
+        options: vec![],
+    };
+    let test_conn = Connection::connect(test_params, TlsMode::None).unwrap();
+    test_server.add_conn(test_conn);
+
+
+    std::process::exit(0);
+
     // Compile SASS (SCSS) to CSS
     println!("Compiling SASS");
     let (styles_dir, main_scss, main_css) = ("styles", "main.scss", "main.css");
@@ -76,25 +67,31 @@ fn main() {
         println!("File found ({})", p.display());
     }
 
+    let rebuild_db = false;
+
     // Connect to acled db
+    let host = "localhost";
+    let uname = "michael";
+    let db_name = "acled";
+
     println!("Opening database connection");
     let params = ConnectParams {
-        target: ConnectTarget::Tcp("localhost".to_owned()),
+        target: ConnectTarget::Tcp(host.to_owned()),
         port: None,
         user: Some(UserInfo {
-            user: "michael".to_owned(),
+            user: uname.to_owned(),
             password: None,
         }),
-        database: Some("acled".to_owned()),
+        database: Some(db_name.to_owned()),
         options: vec![],
     };
     let conn = Connection::connect(params, TlsMode::None).unwrap();
     println!("Connection opened");
 
-    // let conn = Connection::connect("postgresql://michael@localhost/acled", TlsMode::None).unwrap();
-
-    conn.execute("DROP TABLE IF EXISTS event", &[]).unwrap();
-    // conn.execute("DROP TABLE IF EXISTS event_trunc", &[]).unwrap();
+    if rebuild_db {
+        conn.execute("DROP TABLE IF EXISTS event", &[]).unwrap();
+        conn.execute("DROP TABLE IF EXISTS event_trunc", &[]).unwrap();
+    }
 
     conn.execute("CREATE TABLE IF NOT EXISTS event_trunc (\
                     id           SERIAL PRIMARY KEY,\
@@ -121,16 +118,19 @@ fn main() {
 
     let add_event_trunc =
         conn.prepare("INSERT INTO event_trunc (event_date, year, event_type, \
-                      actor_1,ally_actor_1, actor_2, ally_actor_2, country,admin_1, admin_2, \
-                      admin_3, location,latitude, longitude, source, notes,fatalities) VALUES \
-                      ($1, $2, $3, $4,$5, $6, $7, $8,$9, $10, $11, $12,$13, $14, $15, $16,$17)")
+                      actor_1,ally_actor_1, actor_2, ally_actor_2, country,admin_1, \
+                      admin_2, admin_3, location,latitude, longitude, source, \
+                      notes,fatalities) VALUES ($1, $2, $3, $4,$5, $6, $7, $8,$9, $10, \
+                      $11, $12,$13, $14, $15, $16,$17)")
             .unwrap();
 
-    let get_event_trunc_country = conn.prepare("SELECT * FROM event_trunc WHERE country = $1")
-        .unwrap();
+    let get_event_trunc_country =
+        conn.prepare("SELECT * FROM event_trunc WHERE country = $1")
+            .unwrap();
 
     let get_event_trunc_country_year =
-        conn.prepare("SELECT * FROM event_trunc WHERE country = $1 AND year = $2").unwrap();
+        conn.prepare("SELECT * FROM event_trunc WHERE country = $1 AND year = $2")
+            .unwrap();
 
 
     conn.execute("CREATE TABLE IF NOT EXISTS ctry_pair (\
@@ -141,8 +141,10 @@ fn main() {
                  &[])
         .unwrap();
 
-    let add_ctry_pair = conn.prepare("INSERT INTO ctry_pair (name, link) VALUES ($1, $2)").unwrap();
-    let get_all_ctry_pair = conn.prepare("SELECT * FROM  ctry_pair ORDER BY name").unwrap();
+    let add_ctry_pair = conn.prepare("INSERT INTO ctry_pair (name, link) VALUES ($1, $2)")
+        .unwrap();
+    let get_all_ctry_pair = conn.prepare("SELECT * FROM  ctry_pair ORDER BY name")
+        .unwrap();
 
     let mut cur_rows: i64 = 0;
     for cnt in &conn.query("SELECT count(*) from event_trunc", &[]).unwrap() {
@@ -156,12 +158,8 @@ fn main() {
     let country_index: usize = 14;
     let fatality_index: usize = 24;
 
-    let mut country_nl_vec: Vec<CountryNameLink> = Vec::new();
-    let mut country_vec: Vec<Country> = Vec::new();
     let mut all_ctries = AllCountries::new();
     let mut tot_eve: i32 = 0;
-    let mut last_country: Option<String> = None;
-
     let mut rows_aff = 0;
     for row in rdr.records() {
         tot_eve += 1;
@@ -194,47 +192,23 @@ fn main() {
         let country = row[country_index].parse::<String>().unwrap();
         let fatalities = row[fatality_index].parse::<i32>().unwrap();
 
-        all_ctries.push_new_if_not(country.clone());
+        all_ctries.push_new_if_not(country);
 
         let ac_ind = all_ctries.countries.len() - 1;
         all_ctries.countries[ac_ind].num_events += 1;
         all_ctries.countries[ac_ind].num_fatalities += fatalities;
         all_ctries.countries[ac_ind].events.push(Event::from_row(&row));
-
-        if country_vec.is_empty() || last_country.is_none() {
-            country_nl_vec.push(CountryNameLink::new(country.clone()));
-        }
-        if last_country.is_some() && last_country.unwrap() != country {
-            country_nl_vec.push(CountryNameLink::new(country.clone()));
-        }
-
-        last_country = Some(country.clone());
-
-        if country_vec.is_empty() || !contains_name(&country_vec, country.clone()) {
-            let link = country.replace(" ", "");
-            let n_ctry = Country::new(link, country.clone(), 0, 0);
-            country_vec.push(n_ctry.clone());
-        }
-        let ind = country_vec.len() - 1;
-        country_vec[ind].num_events += 1;
-        country_vec[ind].num_fatalities += fatalities;
-        country_vec[ind].events.push(Event::from_row(&row));
     }
     println!("Rows Affected: {}", rows_aff);
     println!("Total Events: {}", tot_eve);
 
-    for elem in &all_ctries.countries {
-        println!("{}, {}", elem.name, elem.num_events);
-    }
-
-    let mut cur_rows: i64 = 0;
-    for cnt in &conn.query("SELECT count(*) from ctry_pair", &[]).unwrap() {
-        cur_rows = cnt.get(0);
-    }
-    if cur_rows == 0 {
-        for elem in country_nl_vec.clone() {
-            let _ = add_ctry_pair.execute(&[&elem.name, &elem.link]).unwrap();
-        }
+    let mut nl_vec: Vec<CountryNameLink> = Vec::new();
+    for elem in &get_all_ctry_pair.query(&[]).unwrap() {
+        let cnv = CountryNameLink {
+            name: elem.get(1),
+            link: elem.get(2),
+        };
+        nl_vec.push(cnv);
     }
 
     let mut server = Nickel::new();
@@ -245,32 +219,56 @@ fn main() {
     server.mount("/js/", StaticFilesHandler::new("src/js"));
 
     // Main Page
-    let main_page = MainPageData { countries: country_nl_vec };
+    let main_page = MainPageData { countries: nl_vec };
     server.get("/",
                middleware! { |request, response| 
         println!("{}", request.path_without_query().unwrap());
     	return response.render("views/main.tpl", &main_page);
     });
     // Page for a country
+    let t_all_ctries = all_ctries.clone();
     server.get("/country/:country_link",
                middleware! { |request, response| 
         println!("{}", request.path_without_query().unwrap());
-        let c_link = request.param("country_link").unwrap();
-        let ctry = get_country_by_link(&country_vec, c_link.to_string());
-        if ctry.is_none() {
-            let c_struct = CountryPageData {
-                found: false,
-                name: "Not Found".to_string(),
-                link: "NotFound".to_string(),
-                years: Vec::new(),
-            };
-            return response.render("views/country.tpl", &c_struct);
+        let link = request.param("country_link").unwrap_or("");
+        let ctry = t_all_ctries.get_by_link(link.to_string());
+        match ctry {
+            Some(c) => {
+                let data = c.to_page_data();
+                return response.render("views/country.tpl", &data);
+            },
+            None => {
+                let data = CountryPageData::new("Not Found".to_string());
+                return response.render("views/country.tpl", &data);
+            },
         }
-        else {
-            let ctry = ctry.unwrap();
-            let pg_data = ctry.to_page_data();
-            return response.render("views/country.tpl", &pg_data);
+    });
+    // Page for a year for a country
+    let t_all_ctries = all_ctries.clone();
+    server.get("/year/:country_link/:year",
+               middleware!{ |request, response|
+        println!("{}", request.path_without_query().unwrap());
+        if request.query().get("sort").is_some() {
+            println!("got sort");
         }
+        let link = request.param("country_link").unwrap();
+        let year = request.param("year").unwrap().parse::<i32>().unwrap();
+        let ctry = t_all_ctries.get_by_link(link.to_string());
+        match ctry {
+            Some(c) => {
+                let data = c.to_year_data(year);
+                return response.render("views/year.tpl", &data);
+            }
+            None => {
+                let data = YearPageData {
+                    eve_vec: Vec::new(),
+                    name: "".to_owned(),
+                    year: 0i32,
+                };
+                return response.render("views/year.tpl", &data);
+            }
+        }
+
     });
     // return JSON data for a country (for xjax)
     // server.get("/country/data/:country_link", middleware! {|request, response|
@@ -278,10 +276,7 @@ fn main() {
     // });
     //
 
-    // Page for a year for a country
-    // server.get("/country/:country_link/:year", middleware!{|request, response|
-    //
-    // });
+
 
 
     server.listen("127.0.0.1:6767");
